refactoring add class Smartphone and LawnGrass from class Product edit Product.__add__
