+++ src/category.py
@@ -1,0 +1,33 @@
+from src.product import *
+
+
+class Category:
+    name: str
+    description: str
+    __products: [Product]
+    category_count = 0
+    product_count = 0
+
+    def __init__(self, name, description, products):
+        self.name = name
+        self.description = description
+        self.__products = products
+        Category.category_count += 1
+        Category.product_count += len(products)
+
+    def __str__(self):
+        sum_products_quantity = sum([product.quantity for product in self.__products])
+        return f"{self.name}, количество продуктов: {sum_products_quantity} шт."
+
+    def add_product(self, product: Product):
+        if isinstance(product, Product):
+            self.__products.append(product)
+            Category.product_count += 1
+
+    @property
+    def products(self):
+        result = ""
+        for product in self.__products:
+            result += f"{product}\n"
+
+        return result
